Remove zod from client-utils
