+++ packages/libs/client-utils/src/built-in/describe-module.ts
@@ -1,25 +1,19 @@
 import { execution } from '@kadena/client/fp';
 import { pipe } from 'ramda';
-import { z } from 'zod';
 import { dirtyReadClient } from '../core/client-helpers';
 import type { IClientConfig } from '../core/utils/helpers';
 
 /**
  * @alpha
  */
-export const describeModuleSchema = z.object({
-  hash: z.string().optional(),
-  blessed: z.array(z.string()).optional(),
-  keyset: z.string().optional(),
-  interfaces: z.array(z.string()).optional(),
-  name: z.string(),
-  code: z.string(),
-});
-
-/**
- * @alpha
- */
-export type DescribeModuleOutput = z.infer<typeof describeModuleSchema>;
+export interface IDescribeModuleOutput {
+  hash: string;
+  blessed: string[];
+  keyset: string;
+  interfaces: string[];
+  name: string;
+  code: string;
+}
 
 /**
  * @alpha
@@ -27,16 +21,11 @@ export type DescribeModuleOutput = z.infer<typeof describeModuleSchema>;
 export const describeModule = async (
   module: string,
   config: Omit<IClientConfig, 'sign'>,
-): Promise<DescribeModuleOutput> => {
+): Promise<IDescribeModuleOutput> => {
   const command = pipe(
     () => `(describe-module "${module}")`,
     execution,
     dirtyReadClient(config),
   );
-
-  const executed = await command().execute();
-
-  const parsed = describeModuleSchema.parse(executed);
-
-  return parsed;
+  return command().execute() as unknown as IDescribeModuleOutput;
 };
